+++ dwds/lib/dwds.dart
@@ -29,4 +29,4 @@ export 'src/services/expression_compiler.dart'
 export 'src/services/expression_compiler_service.dart'
     show ExpressionCompilerService;
 export 'src/utilities/sdk_configuration.dart'
-    show SdkConfiguration, SdkConfigurationProvider;
+    show SdkLayout, SdkConfiguration, SdkConfigurationProvider;