+++ webdev/CHANGELOG.md
@@ -1,10 +1,16 @@
-## 2.7.13-dev
+## 3.0.0-dev
 
 - Add `--enable-experiment` flag to webdev commands and pass it
   to the build runner and the expression compiler service.
-- Prepare or Dart 3 alpha breaking changes:
-  - Move weak null safety tests to special branch of `build_web_compilers`.
+
+**Breaking changes**
+
+- Support Dart 3.0-alpha breaking changes:
   - Do not pass `--(no)-sound-null-safety` flag to build daemon.
+  - Update `build_web_compilers` constraint to `^4.0.0`.
+  - Update `build_runner` constraint to `^2.4.0`.
+  - Update min dart SDK constraint to `3.0.0-134.0.dev`.
+  - Support changes in the SDK layout for Dart 3.0.
 
 ## 2.7.12